refactor(app): add hemoji class to emoji in title
Refactor the main app component to have the correct class applied to the
title emoji. As I'm not that good at regex (I googled this regex), it
has a capture group for emoji at the beginning of the title. It'd be
awesome if someone actually good at regex could improve this to detect
multiple emoji and emoji in different positions of the title.

Closes #30