Fix npm and default run command in Docker container
The package-lock.json should be included in the container for `npm install` time to prevent
version mismatches when using tilde or carrot package.json dependency declarations.

Additionally, adding a default command will allow the container to be built, and then simply run with `docker run` instead of currently depending on using the docker-compose file.

It can be useful to not require docker-compose if a user simply wants to run a pre-built and published container, and has no desire to mount volumes or alter ports. For example:

`docker run -p 3000:3000 electric-io:1.0.1`