feature: rewrite error handling
In order to allow the frontend to consume backend errors in straightforward manner, I rewrote most of the code related to sending API responses via express and the API wrappers in the frontend.

Requests resulting in a non-success status code will now always return a response body of the following structure:

```
{
  data: {
    message: string
  }
}
```

If this structure is present, the API wrappers in the frontend will throw `data.message` as the content of a new error. Otherwise, the response’s status text (e.g. “Internal Server Error”) is used.

In places where that’s possible, the frontend will now notice these error messages and show a toast notification. For example, if for some reason, the call to `fs.writeFile` fails upon trying to save the dashboard settings, a notification reading “Could not write to dashboard.json” will be created. Likewise, if the dashboard.json file is not valid JSON, the dashboard won’t load, but now, there will be a toast notification letting the user know (and it won’t automatically be dismissed).

Also, async/await is beautiful!