Fix card settings not always being saved correctly (#202)
Fixes a bug where the current content of a card settings form was lost on certain actions (e.g. clicking outside the card or clicking the save button). This made it look like the data wasn’t being saved. What actually happened was the form data being reset just when save was clicked.

Adds an early return to the callback processing socket messages. It currently throws on a property systemProperties being accessed when it’s not always defined.