Add ElectricToaster component (#203)
* feature: add ElectricToaster component

Adds a new component called ElectricToaster. It’s a smol toast notification component with the following features:

- Toasts are provided as an array of objects. The parent component (in electric-io, that’s App.vue) is responsible for providing the toast objects.
- Toasts can be created with the properties “content”, “shouldAutoDismiss”, and “autoDismissTimeoutInSeconds”.
  - `content`: Required. This property must contain the content of the toast as a string.
  - `shouldAutoDismiss`: Optional. Controls whether this toast will be automatically dismissed. When not provided, the ElectricToaster’s default will be used (currently set to 10)
  - `autoDismissTimeoutInSeconds`: Optional. Controls the timeout after which a toast will be dismissed. If `shouldAutoDismiss` is `false`, this property has no effect.
- Active (i.e. not dismissed) toast messages are automatically placed in a visually-hidden ARIA live region with the attributes `role="region"` and `aria-live="polite"`. This is intended to cause screen readers to announce the content of the live region as soon as any in-progress announcements are finished.

* feature: rewrite error handling

In order to allow the frontend to consume backend errors in straightforward manner, I rewrote most of the code related to sending API responses via express and the API wrappers in the frontend.

Requests resulting in a non-success status code will now always return a response body of the following structure:

```
{
  data: {
    message: string
  }
}
```

If this structure is present, the API wrappers in the frontend will throw `data.message` as the content of a new error. Otherwise, the response’s status text (e.g. “Internal Server Error”) is used.

In places where that’s possible, the frontend will now notice these error messages and show a toast notification. For example, if for some reason, the call to `fs.writeFile` fails upon trying to save the dashboard settings, a notification reading “Could not write to dashboard.json” will be created. Likewise, if the dashboard.json file is not valid JSON, the dashboard won’t load, but now, there will be a toast notification letting the user know (and it won’t automatically be dismissed).

Also, async/await is beautiful!