Ignore the user's .env file when running tests
This change ignores the configuration in the user's .env file when
running tests (that is, when NODE_ENV === "test"). This fixes issue #210
where the route tests fail if simulation mode is disabled in the user's
.env file.

Currently, the express route tests (lib/routes.test.js) do not pass if
simulation mode is disabled in the user's .env file. This is because
they launch the HTTP server and the hub service and the live hub service
fails to start.

The reason this change fixes the route tests is because it forces the
use of the simulation hub by ignoring the .env file.

Fixing the tests to pass even when using the live hub could be an
enhancement for the future.