chore(deps): update some dependencies
Updates some dependencies in the project. All vue-related dependencies were updated to versions not introducing breaking changes. All eslint-related dependencies were updated as far as possible.

Changes the husky pre-commit hook to just run the eslint command with the `--fix` argument and drops the pretty-quick and prettier packages. They’ve always caused quite a bit of friction and the occassional conflict with the eslint configuration. All of the explicitly configured eslint rules were also removed in an attempt to make the whole topic of linting less stressful and annoying. Let’s see how this pans out.